Fix side panel email reply crashing with Invalid Configuration
PageLayoutRecordPageRenderer subscribed to sidePanelWidgetFooterCommandMenuItemsState
and rendered the widget tree that writes it. EmailThreadComposerFooterEffect writes
that atom from inside the composer, so every write re-rendered the composer, which
scheduled the next write: an unbounded render loop React surfaced as a message-less
AggregateError, reported by the widget as Invalid Configuration. Opening the reply
composer is what starts it, because the composer is what publishes footer actions.

The footer's other two actions already subscribe to that atom themselves, from
inside the footer where they are siblings of the widget tree rather than ancestors.
Move the widget-owned buttons to the same place, so the record page renderer holds
no subscription and the write can no longer reach the writer.
